backend/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
psycopg2-binary
python-dateutil
pydantic
